backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# ── LangGraph / LangChain ─────────────────────────────────────────────────────
langgraph==0.2.60
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-core==0.3.28
langchain-community==0.3.13

# ── Anthropic SDK ─────────────────────────────────────────────────────────────
anthropic==0.42.0

# ── Web search & scraping ─────────────────────────────────────────────────────
tavily-python==0.5.0
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1

# ── Data sources ──────────────────────────────────────────────────────────────
newsapi-python==0.2.7
feedparser==6.0.11
# alpha_vantage requests are made via httpx directly

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── AWS / S3 ──────────────────────────────────────────────────────────────────
boto3==1.35.90
aioboto3==13.2.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.20
structlog==24.4.0
tenacity==9.0.0
aiofiles==24.1.0
orjson==3.10.12
